issue/3 - change Integer to Long for pageID's.

+++ swagger-confluence-cli/src/main/java/net/slkdev/swagger/confluence/cli/SwaggerConfluence.java
@@ -98,13 +98,13 @@ public class SwaggerConfluence {
     private static SwaggerConfluenceConfig buildSwaggerConfluenceConfig(final CommandLine commandLine){
         final SwaggerConfluenceConfig swaggerConfluenceConfig = new SwaggerConfluenceConfig();
         final String ancestorIdString = commandLine.getOptionValue("a");
-        final Integer ancestorId;
+        final Long ancestorId;
 
         if(ancestorIdString == null){
             ancestorId = null;
         }
         else {
-            ancestorId = Integer.valueOf(ancestorIdString);
+            ancestorId = Long.valueOf(ancestorIdString);
         }
 
         swaggerConfluenceConfig.setAncestorId(ancestorId);
